Pause and MainGame menu are working smoothly now (See description below)
Currently, if we hit pause (main game canvas) then it will show Pause canvas which includes resume, menu and setting button. If we click on menu or setting, it will show up those corresponding canvas, and now it will have an extra button, which is a resume button from setting or from main menu. If in main menu, and we click Play instead of resume, the game will start all over again!